reduce rerenders + contain strict
