--- packages/frosted-ui/src/forked-primitives/avatar.tsx
@@ -1,200 +1,0 @@
-import { createContextScope } from '@radix-ui/react-context';
-import { Primitive } from '@radix-ui/react-primitive';
-import { useCallbackRef } from '@radix-ui/react-use-callback-ref';
-import { useLayoutEffect } from '@radix-ui/react-use-layout-effect';
-import * as React from 'react';
-
-import type { Scope } from '@radix-ui/react-context';
-
-/* -------------------------------------------------------------------------------------------------
- * Avatar
- * -----------------------------------------------------------------------------------------------*/
-
-const AVATAR_NAME = 'Avatar';
-
-type ScopedProps<P> = P & { __scopeAvatar?: Scope };
-const [createAvatarContext, createAvatarScope] = createContextScope(AVATAR_NAME);
-
-type ImageLoadingStatus = 'idle' | 'loading' | 'loaded' | 'error';
-
-type AvatarContextValue = {
-  imageLoadingStatus: ImageLoadingStatus;
-  onImageLoadingStatusChange(status: ImageLoadingStatus): void;
-};
-
-const [AvatarProvider, useAvatarContext] = createAvatarContext<AvatarContextValue>(AVATAR_NAME);
-
-type AvatarElement = React.ElementRef<typeof Primitive.span>;
-type PrimitiveSpanProps = React.ComponentPropsWithoutRef<typeof Primitive.span>;
-interface AvatarProps extends PrimitiveSpanProps {}
-
-const Avatar = React.forwardRef<AvatarElement, AvatarProps>((props: ScopedProps<AvatarProps>, forwardedRef) => {
-  const { __scopeAvatar, ...avatarProps } = props;
-  const [imageLoadingStatus, setImageLoadingStatus] = React.useState<ImageLoadingStatus>('idle');
-  return (
-    <AvatarProvider
-      scope={__scopeAvatar}
-      imageLoadingStatus={imageLoadingStatus}
-      onImageLoadingStatusChange={setImageLoadingStatus}
-    >
-      <Primitive.span {...avatarProps} data-status={imageLoadingStatus} ref={forwardedRef} />
-    </AvatarProvider>
-  );
-});
-
-Avatar.displayName = AVATAR_NAME;
-
-/* -------------------------------------------------------------------------------------------------
- * AvatarImage
- * -----------------------------------------------------------------------------------------------*/
-
-const IMAGE_NAME = 'AvatarImage';
-
-type AvatarImageElement = React.ElementRef<typeof Primitive.img>;
-type PrimitiveImageProps = React.ComponentPropsWithoutRef<typeof Primitive.img>;
-interface AvatarImageProps extends PrimitiveImageProps {
-  onLoadingStatusChange?: (status: ImageLoadingStatus) => void;
-}
-
-const AvatarImage = React.forwardRef<AvatarImageElement, AvatarImageProps>(
-  (props: ScopedProps<AvatarImageProps>, forwardedRef) => {
-    // eslint-disable-next-line @typescript-eslint/no-empty-function
-    const { __scopeAvatar, src, onLoadingStatusChange = () => {}, ...imageProps } = props;
-    const context = useAvatarContext(IMAGE_NAME, __scopeAvatar);
-    const imageLoadingStatus = useImageLoadingStatus(src, imageProps.referrerPolicy);
-    const handleLoadingStatusChange = useCallbackRef((status: ImageLoadingStatus) => {
-      onLoadingStatusChange(status);
-      context.onImageLoadingStatusChange(status);
-    });
-
-    useLayoutEffect(() => {
-      if (imageLoadingStatus !== 'idle') {
-        handleLoadingStatusChange(imageLoadingStatus);
-      }
-    }, [imageLoadingStatus, handleLoadingStatusChange]);
-
-    return imageLoadingStatus === 'loaded' ? <Primitive.img {...imageProps} ref={forwardedRef} src={src} /> : null;
-  },
-);
-
-AvatarImage.displayName = IMAGE_NAME;
-
-/* -------------------------------------------------------------------------------------------------
- * AvatarFallback
- * -----------------------------------------------------------------------------------------------*/
-
-const FALLBACK_NAME = 'AvatarFallback';
-
-type AvatarFallbackElement = React.ElementRef<typeof Primitive.span>;
-interface AvatarFallbackProps extends PrimitiveSpanProps {
-  delayMs?: number;
-}
-
-const AvatarFallback = React.forwardRef<AvatarFallbackElement, AvatarFallbackProps>(
-  (props: ScopedProps<AvatarFallbackProps>, forwardedRef) => {
-    const { __scopeAvatar, delayMs, children, ...fallbackProps } = props;
-    const context = useAvatarContext(FALLBACK_NAME, __scopeAvatar);
-    const [canRender, setCanRender] = React.useState(delayMs === undefined);
-
-    React.useEffect(() => {
-      if (delayMs !== undefined) {
-        const timerId = window.setTimeout(() => setCanRender(true), delayMs);
-        return () => window.clearTimeout(timerId);
-      }
-    }, [delayMs]);
-
-    return canRender && context.imageLoadingStatus !== 'loaded' ? (
-      <Primitive.span {...fallbackProps} ref={forwardedRef}>
-        {context.imageLoadingStatus === 'idle' || context.imageLoadingStatus === 'loading' ? null : children}
-      </Primitive.span>
-    ) : null;
-  },
-);
-
-AvatarFallback.displayName = FALLBACK_NAME;
-
-/* -----------------------------------------------------------------------------------------------*/
-
-function resolveLoadingStatus(image: HTMLImageElement | null, src?: string): ImageLoadingStatus {
-  if (!image) {
-    return 'idle';
-  }
-  if (!src) {
-    return 'error';
-  }
-  if (image.src !== src) {
-    image.src = src;
-  }
-  return image.complete && image.naturalWidth > 0 ? 'loaded' : 'loading';
-}
-
-function useImageLoadingStatus(src?: string, referrerPolicy?: React.HTMLAttributeReferrerPolicy) {
-  const isHydrated = useIsHydrated();
-  const image = React.useRef<HTMLImageElement | null>(null);
-  const img = (() => {
-    if (!isHydrated) return null;
-    if (!image.current) {
-      image.current = new window.Image();
-    }
-    return image.current;
-  })();
-
-  const [loadingStatus, setLoadingStatus] = React.useState<ImageLoadingStatus>(() => resolveLoadingStatus(img, src));
-
-  useLayoutEffect(() => {
-    setLoadingStatus(resolveLoadingStatus(img, src));
-  }, [img, src]);
-
-  useLayoutEffect(() => {
-    const updateStatus = (status: ImageLoadingStatus) => () => {
-      setLoadingStatus(status);
-    };
-
-    if (!img) return;
-
-    const handleLoad = updateStatus('loaded');
-    const handleError = updateStatus('error');
-    img.addEventListener('load', handleLoad);
-    img.addEventListener('error', handleError);
-    if (referrerPolicy) {
-      img.referrerPolicy = referrerPolicy;
-    }
-
-    return () => {
-      img.removeEventListener('load', handleLoad);
-      img.removeEventListener('error', handleError);
-    };
-  }, [img, referrerPolicy]);
-
-  return loadingStatus;
-}
-
-function subscribe() {
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
-  return () => {};
-}
-
-function useIsHydrated() {
-  return React.useSyncExternalStore(
-    subscribe,
-    () => true,
-    () => false,
-  );
-}
-
-const Root = Avatar;
-const Image = AvatarImage;
-const Fallback = AvatarFallback;
-
-export {
-  //
-  Avatar,
-  AvatarFallback,
-  AvatarImage,
-  createAvatarScope,
-  Fallback,
-  Image,
-  //
-  Root,
-};
-export type { AvatarFallbackProps, AvatarImageProps, AvatarProps };
